Split DistributedNavigationDelegateTests to multiple files
